firmware: correct the get_cursor() argument types
We only ever pass uint8_t * to get_cursor(), so update the prototypes
and fix the following compiler warnings,

In function ‘__ansicon_open’:
com32/lib/sys/ansicon_write.c:93:6: warning: passing argument 1 of ‘firmware->o_ops->get_cursor’ from incompatible pointer type [enabled by default]
com32/lib/sys/ansicon_write.c:93:6: note: expected ‘int *’ but argument is of type ‘uint8_t *’
com32/lib/sys/ansicon_write.c:93:6: warning: passing argument 2 of ‘firmware->o_ops->get_cursor’ from incompatible pointer type [enabled by default]
com32/lib/sys/ansicon_write.c:93:6: note: expected ‘int *’ but argument is of type ‘uint8_t *’

+++ com32/include/syslinux/firmware.h
@@ -14,7 +14,7 @@ struct output_ops {
 	void (*beep) (void);
 	void (*get_mode)(int *, int *);
 	void (*set_mode)(uint16_t);
-	void (*get_cursor)(int *, int *);
+	void (*get_cursor)(uint8_t *, uint8_t *);
 };
 
 struct input_ops {
